+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.0.0](https://github.com/Masquerade-Circus/buffalo-bench/compare/0.1.0...1.0.0) (2021-08-09)
+
+
+### ⚠ BREAKING CHANGES
+
+* **main:** Renamed `compare` benchmark field to `compareWith`
+
+### Features
+
+* **main:** add Suite class ([96d595a](https://github.com/Masquerade-Circus/buffalo-bench/commit/96d595a4c7e8eb651738927ebb3e232d6475baa9))
+
 ## 0.1.0 (2021-08-08)